requirements.txt
streamlit
langchain
langchain-huggingface
python-dotenv
pyPDF2